requirements.txt
flask>=2.0.1
torch>=1.9.0
ultralytics
opencv-python
pillow
pandas
werkzeug